asyncio implementation of concurrent piece downloading
Experimental asyncio implementation of concurrent piece downloading. This
supports basic retries and timeouts. However, because the piece download
function is synchronous and is used with asyncio.to_thread, it is not
actually cancellable. Therefore, timeouts do not work as expected: The thread
keeps running in the background rather than being properly cancelled.