Implement timeouts with asyncio TaskGroups
Use asyncio TaskGroups and an event loop to manage concurrent
downloading of pieces. Each piece download uses a @timeout wrapper to
send a signal to terminate pieces that are not making progress; this
raises a TimeoutError which then bubbles up and leads to a retry with a
new random peer.

Previous attempts to implement this with threading, asyncio.timeout,
multiprocessing, and the third-party library pebble all failed, because
they did not correctly combine concurrent function calls with the
correct timeout able to terminate them fully. This implementation works,
because it uses system signals to terminate execution of a function, and
uses a single-threaded event loop so threading is not impermissibly
nested.