Refactor to use Torrent class
Encapsulate torrent information in Torrent class. This simplifies
function parameters. It also allows using a reusable GetPeers() method
to request more peers from the tracker if needed.

While doing this refactor, it was discovered that there is a potential
race condition between coroutines when modifying the peer_list and
peer_heap arrays. Use a lock to protect these.