Add semaphore, cache peer connections, use logging
Pairing session: Added a semaphore to restrict concurrent async
connections. Use a cache to store peer connections. Use logging module
to allow more granular logging.

We are also banning bad peers more aggressively, but we discovered
good peers are being banned. The cause of this needs further
investigation, which should be easier with the logging refactor.