Remove unneeded timeout code, improve exception handling
Legacy asyncio timeout code was causing read races. The exact cause
of these races is still under investigation, but they have been
significantly reduced with the removal of this code. They are suspected
to be connected to a race between a reader and the timeout cancellation
of that reader. See https://github.com/encode/httpx/pull/383, for example.

Also handle other named exceptions.

This code is now able to successfully download a complete torrent file.