Implement basic Svelte frontend
Use FastAPI and a TorrentManager instance to allow uploading torrent
files and give a status object on the current download. Svelte frontend
polls the backend once a second to keep things updated. No CSS is
implemented at this time, nor is the important featue of stopping
torrents. (And error propagation to stop the polling also is not working
at this time.)