Frontend: CSS, Progress bar
Add basic CSS styles to frontend. Begin to implement an SVG progress
bar. The goal is to build up the rectangle with verticle bars as each
piece completes. This will illustrate how pieces complete out of order.
However, it is not working yet. There appears to be an overflow problem
in the bottom in the CSS, and the actual updating does not seem to work
properly.

The use of <label> for a single upload torrent is also not very
accessible, we need to look into whether there's a better option.