Fix single-click Select File & Upload Torrent button
Instead of using non-accessible clickable labels, use a button. To hook
the button up with the hidden input element, bind the input to a
variable, add a handler to the button to click the input element, and
add an onchange to the input element which gets the files from the input
element and posts to the API.

The input element needs to take an onchange, not onclick, otherwise it
triggers as soon as the dialogue box opens and posts the first selected
file, instead of waiting for the user to choose a file in the dialogue
box.