Fix viewBox width
The viewBox does not behave properly when the logical width is distinct
from the element width. Although this is not what documents online seem
to describe as the behavior of the viewBox, aligning these two widths
makes the status bar start in the correct place and the widths of the
bars correct.