Add Start and Stop buttons to frontend
Display the correct button depending on the current status of the
downloading torrent. Starting and stopping does not currently respect
tracker the tracker "min interval" key, so this should be used with care
to reduce tracker load.

Also display a confirm dialogue if the user tries to add a new torrent
when another torrent is currently loaded.