Fix mypy issues: Type hints, explicit float conversion
Enable mypy linting for this project. Fix a few typing issues,
particularly concerning the initialization of empty data structures.
There is also an explicit conversion of float to int when calculating
the speed of a peer.

bencoder is an old library without typing hints, so disable that for
mypy in pyproject.toml.