Add skeleton of multifile torrent support
Add a flag if a torrent is multifile, and correctly calculate the size
if it is. Use a dispatcher to write pieces to disk with the appropriate
helper function, although the multifile writer is not yet implemented.

There are some design challenges to face here, including whether we
should pre-calculate file/piece offsets, and how to eventually implement
per-file locks instead of a lock on the whole transaction.